Add combination list feature to new comm panel interface
This closely mirrors the old comm panel, except the algorithm for
combining lists is different.  The old code attempted to combine your OR
and AND settings in a meaningful way, based on the ordering of the lists.
However, the user didn't have control of the ordering of lists, and it
wasn't capable of making arbitrary Boolean statements.  In an effort to
acknowledge these shortcomings and make the logic easier to understand,
the new code simply ANDs (intersects) together all of the lists selected
for AND, then ORs (unions) the result with all of the lists selected for
OR.  Comments welcome.