Add history UI to comm panel
This is the initial implementation of a feature that allows admins to pick
and choose from previous e-mails to initialize a new one.  This could help
people remember what e-mails they sent in the past and what information
was contained in those e-mails.

I think there is a lot to improve here and there may still be some bugs,
so please take a look.  Some problems I can think of:
- Large established chapters are going to see a lot of past e-mails come
  up, so it might be good to have more search/filter options.
- The persistent query filters seem to only apply to a particular program,
  so (for example) if you wanted to take an e-mail from Splash 2011 and
  send it again for Splash 2012, you'd have to just copy out the text and
  start a new message request.

Addresses issue #45.