Give registration types a configurable display name
This adds a new field to the RegistrationType model, which can be set in
order to change the appearance of RegistrationTypes (when rendered using
__unicode__()) throughout the Web site without affecting their behavior.
By default the displayName field is empty, and whenever it is non-empty
the displayName field will be shown instead of the name.

Contains a compilation of changes from pull request #358 -- thanks Ruthie.