Make customforms creation tolerant of missing 'required' field
We were getting errors on some submissions of /customforms/create because
the 'required' field of the metadata was missing.  This change makes the
POST handling code assume that 'required' came from a checkbox (which it
currently does) and make the field non-required unless the 'required'
field is present and set to 'checked'.