Fix server error on some comm panel queries
MIT noticed that when asking for a list of teachers for Spark 2012, you
would get a server error after clicking "Continue to compose e-mail."
This was caused by a string encoding issue in the PersistentQueryFilter
model.  The Q object constructed for the query (e.g. "teachers teaching an
approved class") contained information about the class subjects, including
their Unicode class descriptions.  When this Q object was pickled, the
result was an ASCII string containing some Unicode data.  PostgreSQL (or
any Python string codec) was not happy with this formatting.

The solution is to force the queries to avoid including any fields on
ClassSubject that can contain non-ASCII data and hence trip up
PersistentQueryFilter like that.  A more general solution making
PersistentQueryFilter deal with arbitrary Q objects would be welcome.