Add 'select from checklist' buttons to comm panel
The old comm panel had a button called "I want a subset of this list"
which took you to a page with a checkbox for each user in the QuerySet you
had generated.  The redesigned comm panel UI was lacking such an option,
so I added it.  For the time being it uses the old user search code, but
I think the UI is acceptable.  In the future we should centralize all of
the code that is used for user searching.