Add directions for editing resource type choices
We don't have this capability in any of our views right now, but it is
possible to do with the Django admin interface, so I edited the template
where you would expect to find this feature with an explanatory note.  We
should eventually add this capability to the resource types form where it
belongs, see #372.