Adjust RegistrationType.__unicode__() to avoid test errors
RegistrationTypes display their displayName if it is not empty, and
their name if it is.  We were checking displayName against the
empty string, but sometimes (as is the case during testing) it is None.
This was confusing Django's template renderer because __unicode__() is
supposed to return a Unicode object but was returning None.