Update printables tests for new user selection interface
Recently I made the printables pages use the new comm panel interface
to select the desired list of users, but I neglected to update the
tests to submit the slightly different POST data needed by this interface
to select the same users.  We may have this problem in other places too.