Use from_here in the teacher bio display.
We ask "If no, what is your school/employer?" on the profile form.
So people just don't put down their school if they checked the
from_here box.

We may want to modify userview.html as well, if it's determined that
we care more about what college they go to than exactly what went in
the "College" box.