Avoid using CSS float clearing in program management page
The main program management page (module directory) had a hardcoded
template making use of the CSS "clear: both" property to create a hard
newline after the pretty buttons (due to the way jQuery UI renders them,
you can't make a newline with a <br> tag).  This broke Stanford's layout
and potentially others since they were using CSS floats for their
navigational elements, which were being cleared.  Everything after a
line break had to appear below the bottom of the navigation divs, creating
a huge empty space.

In this commit the divs with the "clear: both" property and removed and
instead each block that needs to appear vertically separate are wrapped in
a div with the "display: inline-block" property.