Add Bootstrap styling to profile forms
I wasn't able to avoid changes to the Python code (which are minor,
and could be substituted using the django-widget-tweaks app); mostly
the templates in the users/profiles directory have changed.  The
previous table-based layout has been replaced with the Bootstrap
recommended form layout, in "horizontal" form so the labels are on
the same row as the fields to save space.