Apply Bootstrap based styling to catalog and fillslot views
The catalog used to be a series of tables.  Now it's a series of divs.
Some custom CSS (fillslot.css and catalog.css consolidated in
catalog.css) is necessary for this, and I haven't yet made this
parameterizable, but that should be possible using the general
LESS generation approach that Vikrant came up with.