Add content_title to template for /medliab
This causes "Splash! 2012 Student Registration" to be displayed in the
title header above the navbar. As far as I know few or none of the
other program modules put anything there, but I think having it looks
better and makes it less confusing for students who expect to be taken
to studentreg.