Move logic of fetch_student_apps into a custom Manager object
Basically, make FormstackStudentApp.objects.fetch(program) do what
FormstackAppSettings.fetch_student_apps() used to do, which was hit
the Formstack API, get the submissions, and update the database
against it.

This makes it possible to have queries on the FormstackStudentApp
table automatically call fetch() so that the database stays up to
date.