Adds getBooleanTag calls to new user registration
The use of getTag was causing problems in some cases.

Some lines had code such as
    Tag.getTag("ask_about_duplicate_accounts",default="false")=="false"
rather than
    Tag.getTag("ask_about_duplicate_accounts",default="false").lower()=="false"
which was causing if-statements to branch the wrong way if the filtered
Tag had a value of "False".

This was noticed when submitting a SinglePhaseUserRegForm with
validation errors caused a UserRegForm to be rendered on page load,
rather than a new SinglePhaseUserRegForm.