Add option for hidng closed sections on class change grid
This is a local change from MIT Splash 2012.  The 'manage class' page can
be used to disable registration for a particular class section without
cancelling or notifying its students.  This causes the registration_status
field on the ClassSection object to be set to 10; the default value of 0
means that registration functions normally.  Previously, all sections were
showing up in the onsite class change grid, even if their registration was
closed; however, trying to add a student to one of those sections failed.
This commit adds an option to show/hide those sections, and they are
hidden by default to avoid confusion.