Add selectable sort criteria to dashboard class list
This adds a dropdown box with some possible criteria like "Name", "ID",
and "Status"; changing the selection causes the sorttable_customkey
attributes of all the table rows to be rewritten, so that the next time
you click on the column header sorttable can order the classes the way you
want them to be ordered.

The "Size" and "Special Requests" options are crudely implemented but in
this case I felt something would be better than nothing.  Please let me
know if you think they should be taken out.