Add role option to ESPUser.addAvailableTime()
I noticed some UserAvailability objects getting created with the Volunteer
role instead of the Teacher role.  To fix that, I made
ESPUser.addAvailableTime() explicitly set the role on each
UserAvailability object it creates. I think we are only using this function
for the Teacher role, so I made that the default.