Send e-mails to teachers when their availability gets forced
I'm not sure how much this feature gets used, but we have a view that sets
the availability of all teachers that have infeasible schedules (not
enough availability to teach the classes they signed up for).  Since this
may still be desired in some circumstances, I made it e-mail the teachers
so they know what's going on.