Hide confirm button on student reg page if confirmation is not allowed
This change checks the Student/Confirm bit and hides the confirm button at
the bottom of the main student reg page if it is not set.  Note that
students can still access the confirmation view by clicking the 'Confirm
Registration' link (in the list of steps at the top of the page) if the
StudentRegConfirm module is included.

Addresses issue #414.