Tweak behavior of AjaxForeignKeyNewformField
If a value was preloaded into the AjaxForeignKeyNewformField, and the user
manually cleared the text (or typed something and failed to select from
the resulting autocomplete options), the underlying value would not be
changed; this isn't consistent with what the user sees.  This happened
even when the initial value was "null".

To make the behavior more intuitive, I changed it so the underlying (e.g.
foreign key) value is cleared whenever the text in the text box gets
changed; it can then be reassigned by selecting from the autocomplete list.

The disadvantage of this is that it's easier to clear the school selected
on the student profile... but hopefully there will be less confusion.