Fix Javascript error in dashboard
The sorttable init code was hitting an error because the 'Loading
classes...' message only occupied one column, whereas the table has a
total of 7 columns.  (It was trying to guess the data type of columns for
which there was no <td> in the table body at the time.)  Making the <td>
wider seems to do the trick.