Remove "remote teacher profile" module
This takes out some code that may have been used for a "Splash on Wheels"
a long time ago, but isn't necessary any more.  Again, the tables are not
being deleted.

I merged the migration from the SAT Prep removal stuff a few commits ago
and gave it a new name, since the migration needed for this was basically
the same.

Addresses issue #361.