Initial commit of survey export.
Exports a dump of survey data from a given program in Excel (2003)
format.  (Note: LibreOffice will read the file, but will take a while to
do it, and will mangle the timestamps.)  The reason for Excel 2003 is
that it's easy to write via templates since it's just XML.  If this
seems like a problem, I can consider modifying it to use a package to
write say XLSX.  CSV isn't really an option since we want to have
multiple tabs, for teach/learn and program/per-class.

Doesn't do the per-class questions yet.