Fix to ClassSubject.teachers data migration
While testing I found that there are sometimes duplicate UserBits
representing the teachers of a class.  This may be because there really
are duplicates (which we didn't care about before) or that a teacher was
removed and then re-added to a class.  This commit takes care of both
those possibilities by using UserBit.valid_objects() and tracking all
(class, teacher) pairs added so far to avoid duplicates.