Fix to UserBit to Record data migration
We were assuming (reasonably) that UserBits recording program actions such
as being confirmed or filling out a survey would be anchored to the
DataTrees corresponding to programs only.  However, it turns out that this
is not the case, probably because of bits created in the admin interface
with incorrect anchors.  This commit excludes bits that are not anchored
to an actual program.