Add models for simpler accounting system
I am attempting to simplify the accounting system that was based in the
accounting_core and accounting_docs apps for the last few years, so that
admins at our chapters can understand it and programs with varying levels
of complexity can all use it.  The new app, accounting, is still based on
the concept of moving money between accounts, but the mechanics are
simpler and it is no longer dependent on the Datatree.

The models are:
- LineItemType, which is a type of item to be purchased
  (This is much like the previous version)
- Account, which represents a pile of money somewhere
- Transfer, which is a movement of money from a 'source' Account to a
  'Destination' account
- FinancialAidGrant, which (when associated with a FinancialAidRequest)
  specifies how a students' costs should be adjusted