Update program creation form for new accounting system
The finaid_cost field is gone because financial aid now works differently.
Rather than specifying a cost for finaid and non-finaid students on each
line item type, you just adjust the amount granted on the FinancialAidGrant.
If we want to restore that behavior, it will be possible to add "rules"
dictating how financial aid is applied to the
IndividualAccountingController.amount_finaid() function.