Remove the FinancialAidRequest.approved field
We don't need this anymore.  A financial aid request is considered
approved if there is a financial aid grant associated with it.  I added an
'approved' property to FinancialAidRequest to maintain the same behavior
when you are looking at a single object, although you can't make a query
based on the 'approved' field anymore.