Make financial aid grants editable on the same page as the request
Financial aid grants can be edited directly from the admin interface, but
this is a much better way.  Now every financial aid request (in the admin
interface) has a "grant" form right below it, where you can enter the
desired dollar amount and/or percentage for the grant.  This substitutes
for writing something in the 'approved', 'amount received', and 'amount
needed' fields that FinancialAidRequest used to have.

Kudos to Django for adding a feature we can use.