Add LineItemOptions model and update other code accordingly
This restores the capability to have an "adjustable" optional cost,
allowing the user to not only select an item but also how much to pay for
it.  The principal use of this is suggesting amounts for donations.

A user's optional cost preferences for a program used to be expressed as a
list of 2-tuples (item name, quantity).  Now it is a list of 3-tuples
(item name, quantity, amount each).  This format is used for the
get_preferences() and apply_preferences() methods of
IndividualAccountingController.