Add functions to represent a user/program accounting record as a string
When someone makes a program payment, we want to know who it's for, and
what program is for.  This bare minimum information is known as the
accounting controller's "ID" and is simply "[program id]/[user id]".

Furthermore, sometimes we'll want to know what specific items the user is
paying for.  The accounting controller's "identifier" includes the ID as
well as a list of line item IDs and amounts paid.  It is possible to check
the consistency of a user's records against one of these strings (for
example, to see if they changed their preferences before their payment was
received), although nothing breaks if they aren't consistent.