Update Cybersource credit card module for new accounting system
The look and feel remain the same, but the identifier that is attached to
the payment form has changed (see previous commit).

Also, the module no longer complains if you have already submitted a
payment.  The effects of previous payments and financial aid grants are
shown on the form and incorporated into the balance.  If you pay for a
program and then select some more optional items, you will be allowed to
pay again.