Update First Data credit card module for new accounting system
These changes are similar to those for the Cybersource module, but I also
took a crack at updating the payment_success view which handles postback
from the payment processor.

Note that I have not tested this with live payments, only verified that
the forms work and produce reasonable POST data to the payment processor.
BOTH CREDIT CARD MODULES MUST BE TESTED ON PRODUCTION SERVERS BEFORE USE.