Add sibling discount support to accounting system
Sibling discounts work much like a financial aid grant: the discount is
calculated based on whether your SplashInfo says so (just like financial
aid is calculated based on the FinancialAidGrant), regardless of what
transfers have already been performed.

For now, the discount is hard coded at $20, but this would be easy enough
to make customizable.  Behavior is unchanged when not using SplashInfo.