Plug SplashInfo into new accounting system
I overrode SplashInfo.save() to set the preferred quantity on 'Saturday
Lunch' and 'Sunday Lunch' line items based on the data contained in the
SplashInfo.  For now, these line item types are hard coded.

Note that the 'splashinfo_choices' and 'splashinfo_costs' Tags must be
properly set in order to use the SplashInfoModule.