Add ProgramAccountingController.remove_pending_transfers()
After the end of a program, you may want to clear out accounting data for
the students that didn't attend (since you don't want to be expecting
thousands of dollars that will never come).  This function makes it
possible to clear the data for any desired set of users (e.g. the ones
that didn't attend).