Use a Tag to provide constant params for template context
This new tag 'template_theme_control' is merely used to store a
theme-dependent configuration structure.  This is so you don't have to
edit template overrides in order to affect the rendering of templates
within a particular theme.

The JSON formatted contents of the tag are dumped into the 'theme' context
variable in render_to_response(), so any elements within the data
structure can be accessed within templates via {{ theme.* }}.