Add theme setup page that can show theme-specific forms
The generic templates process originally had 2 steps: selecting from
available themes, and customizing the appearance of the theme.  Because
the 2nd (and more complex) step is using Vikrant's theme editor and I
don't want to mess with it too much, I decided to just add another step,
which happens in between selecting a them and customizing the appearance.

Why is this necessary?  It's impossible to write truly generic templates
without having some per-site customization of the *content* in addition to
the *appearance*.  There are some things that different chapters are going
to want to do differently, from the way they format the footer to the
organization of navigational links.  Many chapters may not want to edit
the templates directly to impose these changes.

So, our new step collects information from the chapter admins immediately
after they select which theme their site will be using.  Themes can
optionally define a Django form class in config_form.py; if it exists,
then the user can view this form on the /themes/setup/ page.  When they
save the form, the resulting data will be stored in a Tag and show up
as the {{ theme }} variable in template rendering.

There is no need for this form to collect information like colors and
fonts, which are already handled by LESS files and the theme editor, but
they can collect textual and structure information.  A good example would
be the ordering and labeling of the side tabs in the MIT theme.