Add setup form for generic2 theme
Making use of the infrastructure added in recent commits, this change
allows the generic2 theme to be configured.  Configurable settings include
the title and subtitle text, the programs to be featured on the sidebar,
and the structure of the navigation tabs that appear at the top of every
page.  The latter is done using a complex new form field that may be
useful for other themes in some form.