Add 'featured programs' to sidebar
Stanford had the bright idea of putting a link directly to student/teacher
reg for their current program on their sidebar, so that people could get
there in 1 click no matter where they were on the site.  Unfortunately,
this required manually editing the template override for each program.
The idea here is to make such behavior configurable by having a set of
'featured programs' selected on a form.

This is only supported in the generic2 theme right now, but it's possible
that should change.