Add CDN_ADDRESS setting and use it in templates for Bootstrap etc.
This is sort of like Django's MEDIA_URL except for the CDN.  Note that the
CDN should only be used for files that are exactly the same across all of
our sites.  Also, I changed the URL to avoid using our hostname; Amazon
doesn't have the ability to serve our HTTPS certificate, so we just use
their hostname and their certificate.