Fix to teachers' survey/review_single view
For historical reasons, we have survey questions (and answers) that anchor
on ClassSections, and we also have some that anchor on ClassSubjects.
When I made the changes addressing ticket #488, I tested only with my own
survey responses, which (because they were from old programs) were
anchored only on ClassSubjects.  As Gurtej found, this still didn't work
if you had survey responses for a more recent program.  This change
collects survey responses anchored to the combination of their previously
taught sections and previously taught classes, so that there is always
something to see when you click on a survey response link.