Clear Django template loader cache when override is saved
The changes in 6fb43b4 clear the entire TemplateOverride cache when any
template is saved, so that you can modify an override that is included by
another override and your changes will be shown immediately.  However, if
the override you edit is included by a non-overridden template (e.g. in
the filesystem), and the non-overridden template is rendered, Django will
not reload the overridden template and hence does not see your changes.
To fix this problem, we reset Django's template cache every time a
template override is saved.  This is kind of a nuclear option, but I don't
expect it to amount to much of a performance hit in real-world use.

Thanks to Gurtej for pointing this out.