Adds status label to classes in dashboard
Resolves Github Issue #557.

The dashboard color-codes classes by status, but this is not useful for
doing searching. Also, its not obvious what the color codes are unless
you see the tiny message at the top of the dashboard. This change prints
the status very boldly next to the class title, so it is obvious what
the status of the class is, and so it is easily searchable in the
browser.