Fixes student schedule during lottery
d592d67 / Github pull request #434 (the fix for Github Issue #429)
changed StudentClassRegModule.deadline_met() to check for FCFS _or_ LSR
being open. This was a correct behavior for that particular use case of
deadline_met() (to control the display of the "Student registration no
longer open" message on the studentreg main page), but turns out to not
be the correct behavior for all use cases. The change from d592d67 was
causing "add class" links to show up on the studentreg main page during
the lottery phase, which is incorrect behavior.

This change reverts StudentClassRegModule.deadline_met() back to its
original functionality, adds a new method deadline_met_or_lottery_open()
that checks for FCFS or LSR being open, and uses this new method to
control the "Student registration no longer open" message.