Disable context processors when rendering e-mail
In recent versions of Django, Context includes a RequestContext.  A
RequestContext passes your context through the context processor functions
defined in settings.TEMPLATE_CONTEXT_PROCESSORS before rendering.  This is
normally very helpful, but in the case of rendering previews for e-mail
messages it causes a problem.  The 'esp_user' context processor sets the
'user' variable in the context to the user of the current request (e.g.
the admin), overwriting the 'user' variable (ActionHandler object) set by
the comm panel code.  As a result, any smart text placeholders of the form
{{ user.* }} will show the current user, rather than the recipient user.

This commit works around the problem by disabling all of the context
processors while rendering the e-mail preview.
I also added this to the code we use to generate actual e-mail messages
(separately from the comm panel preview), although this may not be
necessary since I don't think the RequestContext is active