Make socket_auth.py remove the socket if present
You can't bind to an AF_UNIX socket if the path already exists. Historically,
we've run socket_auth.py inside of a bash loop that removes the socket and then
starts it. However, in order to run inside of supervisord, it's convenient to
instead have the script itself delete the old socket.