Fix issue #558; allow plus signs in e-mail addresses.
We were using the urllib quote_plus and unquote_plus functions to encode
the e-mail address in between the two steps of user registration forms
(when ask_about_duplicate_accounts = True).  However, this resulted in
plus signs being replaced with spaces, which would then cause the e-mail
address to be incorrect and fail form validation.  Changing back to
quote() and unquote() prevents this, and it's safe because quote() is
called after form.is_valid(), so we know there are no spaces in the e-mail
address to begin with.

Thanks to Manish Sharma for finding the root cause of this issue and
proposing a fix in #579.