Fix problem with classes not showing up in lottery on IE7
For one reason or another there is a <tbody> added inside the <table>
for each timeslot's group of "regular classes" in the lottery interface.
Most browsers allow us to add a <tr> for each class to the end of the
table, but IE7 (and possibly others) require you to add it to the tbody
instead in order for it to show up.