Adjust lottery assignment ordering of sections
This change was suggested by Stanford.  With the sections in completely
random order, the longer classes (e.g. 3--4 hr) tend to get much fewer
students because the students that applied to them have already been
assigned conflicting classes.  This change assigns groups of sections
based on their duration in descending order to reduce the frequency of
this problem.