Stop assuming people were born in 1994 (if using login by bday)
This is a local change from Chicago, and the problem only affected Chicago
because they were the only ones linking to the login by birthday feature.
If you have logged in by birthday and you go to the student profile page,
the code was using the saved month/day from your login and assuming a year
of 1994.  This worked reasonably well (at least in 2008).  However, it's
completely wrong.

We discovered this problem when an unlucky student entered 2/29/1996 (a
valid date during a leap year) on their profile and then proceeded to log
in by birthday and edit their profile again.

Now we seed the DOB field with the previous value.  If none is available,
we will guess 1998 for the time being.