Add option to ajax_sections view to filter out rejected sections
For Ajax scheduling it's desired to fetch a list of all sections that you
actually want to schedule, e.g. those that are accepted and belong to an
accepted subject.  Usually we perform that filtering in the client side.
Now that behavior can be obtained by passing a GET parameter
"accepted_only": /manage/[program]/[instance]/ajax_sections?accepted_only