Fixes bug in create_membership_method
If the user_types Tag is used to define a new user type, that user type
is passed as a unicode string to create_membership_method, and is used
to set the name of a dynamically created method that checks for
membership in this user class. The __name__ attribute of this method
must be a regular string, but it was being cast to a unicode string if
user_class was a unicode string from the user_types Tag.

This fix casts user_class back to a regular string to prevent this
problem.