Change method of creating FormstackStudentProgramApp objects
Instead of fetching the API every time we make a query for
applications, use webhooks and signals to create them as they are
submitted.

The fetch() method still exists for manually resyncing with the API.

Also fixes a bug with the get_subject mapping of strings to
ClassSubjects; the syntax "except a, b" was incorrect.