Add app_is_open flag to control /studentapp
We could use meets_deadline, but the error message doesn't fit very
well with Junction, and userbits are evil anyway. The only thing
app_is_open does is add or remove the button at the bottom that takes
you to the student app; the QSD block on the page should be used to
display appropriate text.