Add a setting for "autopopulated fields" for the Formstack app
Allows one to add fields on the app that should be autopopulated via
the admin panel.

Note: this is a potentially dangerous implementation that uses eval()
on database objects. FormstackAppSettings *should* only be editable
via the admin panel, so this should be okay as long as admins exercise
caution editing the field.