Fixes bug in onsite schedule_students
Fixes bug in ProgramModuleObj.deadline_met() caused by the upgrade to
Django 1.4.

During the upgrade to Django 1.4, specifically in
93d46cfb29f03cc821f301cffa77ec205dbf19a2
we changed our auth middleware. Now, request.user is a SimpleLazyObject
that evaluates to an ESPUser, rather than an ESPUser.

Normally, the SimpleLazyObject resolves requests to its attributes by
passing those requests on to the inner ESPUser object. However, when
the __dict__ attribute of request.user refers to the __dict__ attribute
of the SimpleLazyObject itself, rather than that of the inner ESPUser
object. Thus, checking to see if some key exists in the ESPUser's
__dict__ fails, because the check is actually being performed on the
SimpleLazyObject's __dict__. This was causing onsite schedule_students
to not show links for adding and removing classes, since these links
only appear if deadline_met() is True, which only happens if
'onsite_local' is in the ESPUser's __dict__ and has a value of True.

Fixes this by calling isOnsite() (which correctly resolves to the inner
ESPUser) rather than checking for 'onsite_local' manually. isOnsite()
already performs this check, so we should be good.