Make list of classes on teacherreg page update properly
We had a caching bug because the cached inclusion tag
{% render_teacher_list_row %} was accepting 2 additional arguments besides
the class, which made it impossible for cache dependencies (e.g.
depend_on_row, depend_on_cache) to flush the cache for specific classes.
The solution here is to remove the extra arguments.  I was able to restore
the functionality of the friendly_times_with_date argument within the
template tag, but the 'can_edit' isn't going to work (that's a user
dependent variable which is incompatible with this type of caching
anyway).