Fix QSD caching bug
Our render_qsd template tag was accepting an optional argument 'user'.
This meant that it was possible for the cache to not be flushed properly
because the key set being deleted didn't match the value of the 'user'
argument that was originally used when the cached HTML was rendered.  We
aren't using the user anyway because edit permissions are handled by
Javascript, so there are no side effects to removing this.  See also
9ad33326.