Add 'unschedule' option when cancelling classes
When you cancel a class, you might want to unschedule it to free up space
for others.  This is usually desirable well in advance of the program.
However, immediately before or during the program you may want to keep
track of where it had been scheduled, and you have no use for the extra
space.  Now you can choose.