Adds JSON module to all existing programs
Adds a South data migration to add the JSON Data Module to all existing
programs. The JSON Data Module is a dependency for many commonly-used
modules, so it is important that it be enabled for all programs.

The migration deletes all existing JSON ProgramModuleObjs, to avoid an
edge-case where duplicate JSON ProgramModuleObjs cause Django errors.
This is safe, because ProgramModuleObjs are automatically regenerated.