Fixes walk-ins bug in lottery
The lottery uses a ClassCategory's symbol attribute for certain logic,
including figuring out whether a class is a walk-in seminar or not.
However, the open_class_category's name (its category attribute) was
being passed to the template instead.

Fixes this by passing open_class_category to the template as a JSON
string, and referring to its symbol attribute when necessary.

In the future, it might be strictly better to do ClassCategory logic
based on id, rather than symbol, since it is possible for two
ClassCategory objects to have the same symbol (though we haven't run
into problems yet because we usually avoid this).

Used for Stanford Splash Spring 2013 lottery student registration.