Adds several depends_ons and tweaks migrations to get flush working
One of the changes was to actually delete the
modules_satprepadminmoduleinfo table. Commit 64161115 by @pricem
describes why this was not included in migration 0024 initially,
but this was preventing flushing, and at this point I think we can
safely delete this table. Since this is affecting an old migration
we have to replay this migration to actually cause the tables to
delete, so if MIT ESP wants to keep the data around we can just
leave things as they are.